Changed way of sorting
Sorting through a treemap is a problem if dates arent unique so it is not a good option if we use it as a key. So I changed the sorting method and now I use a interface comparator. Also there is a minor change for leap years and now random generated days in Februaru can go to 29.